chore; more fixes
