+++ src/adapters/sqlite/makeDispatcher/WatermelonEventEmitter.native.js
@@ -1,0 +1,7 @@
+import { NativeEventEmitter } from 'react-native'
+
+export const EventType = {
+  CDC: 'SQLITE_UPDATE_HOOK',
+}
+
+export const WatermelonDBEvents = new NativeEventEmitter(DatabaseBridge)
