+++ src/adapters/sqlite/makeDispatcher/index.js
@@ -5,12 +5,11 @@ import { fromPairs } from 'rambdax'
 
 import DatabaseBridge from './node/DatabaseBridge'
 
-import { logger } from '../../../utils/common'
+import { type ConnectionTag, logger } from '../../../utils/common'
 
 import { fromPromise } from '../../../utils/fp/Result'
 
-// Comment out type import instead of removing it completely
-// import type { DispatcherType, SQLiteAdapterOptions, NativeDispatcher } from '../type'
+import type { DispatcherType, SQLiteAdapterOptions, NativeDispatcher } from '../type'
 
 import { syncReturnToResult } from '../common'
 
@@ -37,7 +36,11 @@ const dispatcherMethods = [
   'enableNativeCDC',
 ]
 
-export const makeDispatcher = (type, tag, _dbName) => {
+export const makeDispatcher = (
+  type: DispatcherType,
+  tag: ConnectionTag,
+  _dbName: string,
+): NativeDispatcher => {
   const methods = dispatcherMethods.map(methodName => {
     const name = type === 'synchronous' ? `${methodName}Synchronous` : methodName
 
@@ -45,7 +48,7 @@ export const makeDispatcher = (type, tag, _dbName) => {
       methodName,
       (...args) => {
         // $FlowFixMe
-        const callback = args[args.length - 1]
+        const callback: any => void = args[args.length - 1]
         const otherArgs = args.slice(0, -1)
 
         if (type === 'synchronous') {
@@ -69,7 +72,7 @@ export const makeDispatcher = (type, tag, _dbName) => {
   return dispatcher
 }
 
-export function getDispatcherType(options) {
+export function getDispatcherType(options: SQLiteAdapterOptions): DispatcherType {
   if (options.synchronous) {
     if (DatabaseBridge.initializeSynchronous) {
       return 'synchronous'
@@ -82,32 +85,3 @@ export function getDispatcherType(options) {
 
   return 'asynchronous'
 }
-
-export const EventType = {
-  CDC: 'SQLITE_UPDATE_HOOK',
-}
-
-// Use require to avoid transpilation issues
-const EventEmitter = require('events')
-
-// Create the native event emitter class
-class NativeEventEmitter extends EventEmitter {
-  constructor() {
-    super()
-  }
-
-  addListener(event, callback) {
-    super.addListener(event, callback)
-    this._event = event
-    this.listener = callback
-    return this
-  }
-
-  remove() {
-    super.removeListener(this._event, this.listener)
-  }
-}
-
-// Create and export the events instance - ensure it's initialized before export
-const watermelonEvents = new NativeEventEmitter()
-export const WatermelonDBEvents = watermelonEvents
